// store/client/types.ts
export interface ClientData {
  id?: number;
  title: string;
  fullname: string;
  email: string;
  address: string;
  phone: string;
  company_name: string;
  work_address: string;
  date_of_birth: string;
  job_title: string;
  status: string;
  created?: string;
  updated?: string;
  deleted?: string;
  reason_for_delete?: string;
}